reformat py script output to JSON
