+++ server/src/transcribe/score_parser.js
